refactor: move content from LICTest to EvaluateLICTest and move LIC3 to EvaluateLIC
This commit will move tests for LIC from the deleted LICTest to EvaluateLICTest and LIC3 from Main to EvaluateLIC. This is to clean up after local merge with main
